Partial implementation - I rewrote the file shuffler so that it could be called as a component from GUI5 and the file shuffler manual program. Need to fix the context still so that when the run file shuffler button is pressed, it knows to call the run_file_shulffler_program function of the file-shuffler-view class (FileShufflerView) in ./file-shuffler/file-shuffler-component
